some fix about service

+++ CallMeMaybeAPI/WebApplication1/Controllers/SalarieController.cs
@@ -20,8 +20,17 @@ namespace WebApplication1.Controllers
         [HttpGet("get/all")]
         public async Task<IActionResult> GetAll()
         {
+
             try
             {
+                if (!Request.Headers.TryGetValue("X-App-Identifier", out var appIdentifier))
+                {
+                    return Unauthorized(new { message = "En-tête X-App-Identifier manquant." });
+                }
+                if (appIdentifier != "CallMeMaybe")
+                {
+                    return Unauthorized(new { message = "Accès refusé : X-App-Identifier invalide." });
+                }
                 var data = await (from salarie in _context.Salarie
                                   join service in _context.Service on salarie.idService equals service.id
                                   join site in _context.Site on salarie.idSite equals site.id
@@ -57,6 +66,15 @@ namespace WebApplication1.Controllers
             {
                 return BadRequest("Les données du salarié sont nulles.");
             }
+            if (!Request.Headers.TryGetValue("X-App-Identifier", out var appIdentifier))
+            {
+                return Unauthorized(new { message = "En-tête X-App-Identifier manquant." });
+            }
+            if (appIdentifier != "CallMeMaybe")
+            {
+                return Unauthorized(new { message = "Accès refusé : X-App-Identifier invalide." });
+            }
+
 
             try
             {
@@ -85,6 +103,14 @@ namespace WebApplication1.Controllers
                 if (salarieToDelete == null)
                 {
                     return NotFound($"Le salarié avec l'ID {id} n'existe pas.");
+                }
+                if (!Request.Headers.TryGetValue("X-App-Identifier", out var appIdentifier))
+                {
+                    return Unauthorized(new { message = "En-tête X-App-Identifier manquant." });
+                }
+                if (appIdentifier != "CallMeMaybe")
+                {
+                    return Unauthorized(new { message = "Accès refusé : X-App-Identifier invalide." });
                 }
 
                 _context.Salarie.Remove(salarieToDelete);
@@ -115,6 +141,16 @@ namespace WebApplication1.Controllers
                 {
                     return NotFound($"Le salarié avec l'ID {id} n'existe pas.");
                 }
+
+                if (!Request.Headers.TryGetValue("X-App-Identifier", out var appIdentifier))
+                {
+                    return Unauthorized(new { message = "En-tête X-App-Identifier manquant." });
+                }
+                if (appIdentifier != "CallMeMaybe")
+                {
+                    return Unauthorized(new { message = "Accès refusé : X-App-Identifier invalide." });
+                }
+
 
                 // Mise à jour des champs
                 salarieToUpdate.nom = salarie.nom;
